ml_service/requirements.txt
flask>=2.0.1,<2.3.0
flask-cors>=3.0.10
numpy==1.24.3
scikit-learn>=1.0.0
joblib>=1.0.1
pandas==2.0.3
scapy==2.5.0
python-dotenv>=0.19.0
python-socketio==5.7.2
eventlet==0.33.3
flask-socketio==5.3.4
python-engineio==4.5.1
werkzeug==2.2.3
psutil>=5.9.0